💄 style: refactor create new note modal
